requirements.txt
mysql-replication
elasticsearch